proyecto_python/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
